Add run utility and update run-local script for improved file handling
